Swap all internal usage of FragmentMeta* for tdb_shared_ptr<FragmentMeta>
This is needed to avoid life time issues between the opened `Array`
object and the `FragmentInfo` object. FragmentInfo could have copied the
`FragmentMeta`, however we want to slowly remove all unneeded usages of
raw pointers. This cause the `FragmentMeta` objects themselves did not
need to be raw pointers and using a `tdb_shared_ptr` made sense and
allows us to avoid the copy.